pyproject.toml
[project]
name = "sanaap-interview"
version = "0.1.0"
description = ""
authors = [
    {name = "tahericode",email = "mrtcode2@gmail.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.1,<8.0.0)",
    "django-minio-storage (>=0.5.9,<0.6.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "djangorestframework-simplejwt (>=5.3.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
